requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
# Flask-MySQLdb==2.0.0
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
# mysqlclient==2.2.4
mysql-connector-python==8.0.33
packaging==24.1
pillow==10.3.0
pypng==0.20220715.0
qrcode==7.4.2
setuptools==70.0.0
typing_extensions==4.12.2
Werkzeug==3.0.3
wheel==0.43.0
